% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dataset.R
\name{align_alleles}
\alias{align_alleles}
\title{Align allele sequences across loci}
\usage{
align_alleles(results_summary, derep = TRUE, ...)
}
\arguments{
\item{results_summary}{cross-sample summary data frame.}

\item{derep}{logical; should allele sequences be dereplicated prior to
alignment?  TRUE by default.}

\item{...}{additional arguments passed to \code{\link[msa]{msa}}.}
}
\value{
list of MSA alignment objects, one per locus.
}
\description{
Create a list of alignments across observed alleles, with one alignment per
locus present in the given summary, and by default one entry for each unique
sequence for each locus.
}
