% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sample.R
\name{allele_match}
\alias{allele_match}
\title{Check Sample Data for Potential Allele Matches}
\usage{
allele_match(sample_data, locus_name)
}
\arguments{
\item{sample_data}{data frame of processed data for sample as produced by
\link{analyze_seqs}.}

\item{locus_name}{character name of locus to match against.}
}
\value{
logical vector of entries for potential alleles.
}
\description{
Check the entries in a processed sample data frame for potential matches to a
given locus.
}
