% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_dataset.R
\name{analyze.dataset}
\alias{analyze.dataset}
\title{Analyze all samples in a dataset}
\usage{
analyze.dataset(dataset, locus_attrs, num.cores = max(1,
  as.integer(parallel::detectCores()/2) - 1),
  summary.function = summarize.sample)
}
\arguments{
\item{dataset}{data frame of sample details as produced by
\code{prepare.dataset}.}

\item{locus_attrs}{data frame of locus attributes as produced by
\code{load.locus_attrs}.}

\item{num.cores}{integer number of CPU cores to use in parallel for sample
analysis.}

\item{summary.function}{function to use when summarizing each sample's full
details into the standard attributes.}
}
\value{
list of results
}
\description{
Load all samples for a given dataset and produce a list of processed samples
and a summary data frame showing each sample's summary row-by-row.
}
