% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{apply_config}
\alias{apply_config}
\title{Apply CHIIMP options globally}
\usage{
apply_config(config, keep = TRUE)
}
\arguments{
\item{config}{configuration options data frame, as from \link{load_config}}

\item{keep}{keep existing CHIIMP options, or remove them before setting new
ones?}
}
\description{
\code{apply_config} parses each row in a configuration options data frame and sets
a corresponding option with \code{options(name = value)}.  All options are
prefixed with "chiimp." to avoid colliding with options used in other
libraries.
}
