% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dataset.R
\name{calc_genotype_distance}
\alias{calc_genotype_distance}
\title{Calculate distance between two genotypes}
\usage{
calc_genotype_distance(g1, g2, na_reject = TRUE)
}
\arguments{
\item{g1}{vector for genotype.}

\item{g2}{vector for genotype.}

\item{na_reject}{logical; should NA entries be treated as mismatches?  TRUE
by default.}
}
\value{
numeric distance score for the pair of input genotypes.
}
\description{
Given two vectors representing genotypes, produce a scalar value for the
distance between them.  The vectors should be sorted the same such that the
each two values of each vector represent two alleles of the same locus.  In
this implementation the distance is simply the number of mismatches between
the alleles for each locus, accounting for either order of alleles.
}
