% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{cfg}
\alias{cfg}
\title{Get/set CHIIMP global configuration options}
\usage{
cfg(key, val)
}
\arguments{
\item{key}{name of a CHIIMP configuration option}

\item{val}{value to set for a particular key}
}
\description{
\code{cfg} is a helper function to provide a shortcut for the equivalent
\code{options(chiimp.key = value)}.  With no arguments, all the currently-defined
settings are returned as a list.  With a key given, the corresponding value
is returned.  With both key and value, a new option value is set.
}
