% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_seqs.R
\name{check_length}
\alias{check_length}
\title{Check sequences for expected length}
\usage{
check_length(sample.data, locus_attrs)
}
\arguments{
\item{sample.data}{data frame of processed sequence data.}

\item{locus_attrs}{data frame of attributes for loci to look for.}
}
\value{
logical vector specifying, for each entry, if the sequence is within
the matching locus' expected length range.
}
\description{
Return a logical vector specifying, for each entry, if the sequence is within
the expected length range for the matched locus.
}
