% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sample.R
\name{check_motif}
\alias{check_motif}
\title{Check sequences for STR repeats}
\usage{
check_motif(sample.data, locus_attrs, nrepeats)
}
\arguments{
\item{sample.data}{data frame of processed sample data.}

\item{locus_attrs}{data frame of attributes for loci to look for.}

\item{nrepeats}{number of repeats of each locus" motif to require for a
match.}
}
\value{
logical vector indicating where repeats were observed.
}
\description{
Return a logical vector specifying, for each entry, if the sequence contains
at least \code{nrepeats} perfect repeats of the matching locus' motif.
}
