% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiimp.R
\docType{package}
\name{chiimp-package}
\alias{chiimp}
\alias{chiimp-package}
\title{Analyze Microsatellites}
\description{
Analyze DNA microsatellites in high-throughput sequencing datasets.
}
\author{
\strong{Maintainer}: Jesse Connell \email{ancon@upenn.edu}

}
