% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{CMP}
\alias{CMP}
\title{Complements of IUPAC nucleotide codes}
\format{
An object of class \code{character} of length 14.
}
\usage{
CMP
}
\description{
\code{CMP} is a named character vector mapping each IUPAC DNA nucleotide code to
its complement.  For example, \code{CMP["A"]} is \code{T}.
}
\keyword{datasets}
