% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sample.R
\name{find_artifact}
\alias{find_artifact}
\title{Check for non-stutter PCR artifacts between sample entries}
\usage{
find_artifact(sample.data, locus_attrs, count.ratio_max = 1/3)
}
\arguments{
\item{sample.data}{data frame of processed sample data.}

\item{locus_attrs}{data frame of attributes for loci to look for.}

\item{count.ratio_max}{comparing the currently-checked entry to another
entry, this is the highest ratio of counts where an entry will still be
considered artifactual}
}
\value{
integer vector specifying, for each entry, the row index for another
  entry that may have produced each entry as an artifactual sequence.
}
\description{
Searches a processed STR sample for entries that may be PCR artifacts, other
than stutter, from another entry in the sample.  Potential artifacts are
sequences with counts lower than another sequence by a given ratio and
sequence length within 1 bp of the other sequence.  This only considers
STR-labeled rows and requires a given entry to have counts at most
\code{count.ratio_max} compared to the candidate "source" entry to be
considered an artifact.  Sequence content is not currently considered, just
relative sequence lengths and counts.
}
