% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{find_primer_matches}
\alias{find_primer_matches}
\title{Find primer matches for reads}
\usage{
find_primer_matches(seqs_reads, seqs_primers, max_mismatches = NA)
}
\arguments{
\item{seqs_reads}{character vector of read sequences}

\item{seqs_primers}{character vector of primer sequences}

\item{max_mismatches}{integer number of mismatches allowed when checking
primers against reads, or \code{NA} to check all combinations}
}
\value{
data frame of read and primer index pairs and match details
}
\description{
Returns a data frame of index values for the two input vectors for each
match, along with start and stop positions within each read and a count of
base mismatches.  There may be zero, one, or multiple output rows per input
read.
}
\details{
Indels are not supported.  Partial matches are supported at the 3' end and
are counted as mismatches.

If \code{max_mismatches} is \code{NA}, the best match for every read and primer
combination will be included in the output.  If \code{max_mismatches} is an
integer, at most one row will be provided for each read (for the first
discovered primer at or below that number of mismatches).  In this mode the
primers are checked for perfect matches first and then re-checked in
decreasing order of abundance to find any imperfect matches.  This means that
a search with a maximum mismatch count specified will be much faster than one
without, but could give counterintuitive results for some edge cases.  For
example, with max mismatches of 10 and a read containing primer A at 5
mismatches and primer B at 8 mismatches, either A or B may be matched
depending on the relative abundance of perfect matches found elsewhere. In
practice, with dissimilar primer sequences and a low number of allowed
mismatches, this is unlikely to be a problem.
}
