% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_seqs.R
\name{find_stutter}
\alias{find_stutter}
\title{Check for PCR stutter between sequences}
\usage{
find_stutter(
  sample.data,
  locus_attrs,
  count.ratio_max = cfg("max_stutter_ratio")
)
}
\arguments{
\item{sample.data}{data frame of processed sample data.}

\item{locus_attrs}{data frame of attributes for loci to look for.}

\item{count.ratio_max}{comparing the currently-checked entry to another
entry, this is the highest ratio of counts where an entry will still be
considered stutter.}
}
\value{
integer vector specifying, for each entry, the row index for another
  entry that may have produced each entry as a stutter band.
}
\description{
Searches a processed STR sample for entries that may be artifacts of PCR
stutter from another entry in the sample.  This only considers STR-labeled
rows and requires a given entry to have counts at most \code{count.ratio_max}
compared to the candidate "source" entry to be considered stutter.  Sequence
content is not currently considered, just relative sequence lengths and
counts.
}
