% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsat.R
\name{format_pandoc_args}
\alias{format_pandoc_args}
\title{Create Pandoc Metadata Argument Strings}
\usage{
format_pandoc_args(metadata)
}
\arguments{
\item{metadata}{named list of
\href{http://pandoc.org/MANUAL.html#metadata-blocks}{pandoc metadata
options}.}
}
\value{
list of \code{--metadata} argument strings
}
\description{
Convert named list of pandoc metadata options into command-line
\code{--metadata} argument strings.
}
