% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiimp.R
\name{full_analysis}
\alias{full_analysis}
\title{Perform a full microsatellite analysis}
\usage{
full_analysis(config, dataset = NULL)
}
\arguments{
\item{config}{list of configuration options.  See the summary for
\code{\link{config.defaults}} for more details.}

\item{dataset}{optional custom data frame to override the sample attributes
defined in the config.}
}
\value{
list of results, with the full configuration list included as
  "config."
}
\description{
Given a list of configuration options, run all aspects of a microsatellite
analysis, and save the corresponding output files.
}
\examples{
# Set up a temporary copy of the CHIIMP test data
example_dir <- tempfile()
dir.create(example_dir)
setwd(example_dir)
test_data$write_seqs(test_data$seqs,
                     "str-dataset",
                      "Replicate1-Sample\%s-\%s.fasta")
locus_attrs_path <- system.file("example_locus_attrs.csv",
                                package = "chiimp")
file.copy(locus_attrs_path, "locus_attrs.csv")
# Run the example analysis
config_path <- system.file("example_config.yml", package = "chiimp")
config <- load_config(config_path)
results <- full_analysis(config)

}
