% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{full_locus_match}
\alias{full_locus_match}
\title{Check Sample Data for Potential Allele Matches}
\usage{
full_locus_match(sample_data, locus_name)
}
\arguments{
\item{sample.data}{data frame of processed data for sample as produced by
\code{\link{analyze_sample}}.}

\item{locus.name}{character name of locus to match against.}
}
\value{
logical vector of entries for potential alleles.
}
\description{
Check the entries in a processed sample data frame for potential matches to a
given locus.
}
