% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{handle_primers}
\alias{handle_primers}
\title{Modify reads based on matched primers}
\usage{
handle_primers(
  result,
  locus_attrs,
  primer_action_fwd,
  primer_action_rev,
  reverse_primer_r1
)
}
\arguments{
\item{result}{data frame of read and primer information as from
\link{make_read_primer_table}}

\item{locus_attrs}{data frame of locus attributes}

\item{primer_action_fwd}{keyword for handling forward primers (see
\link{make_read_primer_table})}

\item{primer_action_rev}{keyword for handling reverse primers (see
\link{make_read_primer_table})}

\item{reverse_primer_r1}{is each reverse primer in \code{locus_attrs} given in its
orientation on the forward read?}
}
\value{
modified data frame with \code{Seq} column containing modified sequences
}
\description{
\code{handle_primers} carries out the read modification specified in
\link{make_read_primer_table}.
}
