% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{histogram}
\alias{histogram}
\title{Plot basic histogram of STR sample}
\usage{
histogram(samp, main = "Number of Reads by Sequence Length",
  locus.name = NULL, sample.summary = NULL, cutoff_fraction = 0.05,
  xlim = range(samp$Length))
}
\arguments{
\item{samp}{data frame of dereplicated sequences.}

\item{main}{title of the plot.}

\item{locus.name}{name of the locus to match alleles for.  If unspecified and
not listed in \code{sample.summary} the locus with the highest matched
counts will be used.  To disable the matching entirely use \code{NA}.}

\item{sample.summary}{summary data frame as prepared by
\code{summarize_sample}.  Used to label the called alleles.}

\item{cutoff_fraction}{numeric threshold for the fraction of locus-matching
counts needed to call an allele.  To disable display use \code{NA}.}

\item{xlim}{numeric range for x-axis.}
}
\description{
Given a processed STR sample, plot a histogram of counts per sequence length.
If selected, bars for the counts of sequences matching a selected locus, and
the counts of sequences representing called alleles, will be overlayed.
}
