% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load.locus_attrs}
\alias{load.locus_attrs}
\title{Load table of locus attributes}
\usage{
load.locus_attrs(fp.locus_attrs, ...)
}
\arguments{
\item{fp.locus_attrs}{path to text file}
}
\value{
data frame of locus attributes
}
\description{
Load a tab-separated table of locus attributes to use for analysis.
}
