% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_allele_names}
\alias{load_allele_names}
\title{Load table of known allele names}
\usage{
load_allele_names(fp, ...)
}
\arguments{
\item{fp}{path to text file.}

\item{...}{additional arguments passed to \link{load_csv}.}
}
\value{
data frame of allele names
}
\description{
Load a comma-separated table of allele names for use in reporting.
}
