% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_config}
\alias{load_config}
\title{Load configuration file}
\usage{
load_config(fp)
}
\arguments{
\item{fp}{path to configuration file.}
}
\value{
data frame of configuration options
}
\description{
Load a CSV or YAML-formatted text file of configuration options for
microsatellite analysis.  The \link{main} function loads configuration options
with this and sets them package-wide with \link{apply_config}.
}
\details{
Whatever entries are present in the file will be returned, but names that
don't match known names (see \link{CFG_DEFAULTS}) will be reported in a warning.
}
\examples{
filename <- system.file("example_config.yml", package = "chiimp")
config <- load_config(filename)
# And then: full_analysis(config)
}
