% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_config_csv}
\alias{load_config_csv}
\title{Load configuration file from CSV}
\usage{
load_config_csv(fp)
}
\arguments{
\item{fp}{path to configuration file.}
}
\value{
data frame of configuration information
}
\description{
Load configuration file from CSV
}
