% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_config_yaml}
\alias{load_config_yaml}
\title{Load configuration file from YAML}
\usage{
load_config_yaml(fp)
}
\arguments{
\item{fp}{path to configuration file.}
}
\value{
data frame of configuration information
}
\description{
Load configuration file from YAML
}
