% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_dataset}
\alias{load_dataset}
\title{Load table of sample attributes}
\usage{
load_dataset(fp, ...)
}
\arguments{
\item{fp}{path to text file.}

\item{...}{additional arguments passed to \link{load_csv}.}
}
\value{
data frame of sample attributes for the dataset.
}
\description{
Load a comma-separated table of sample attributes for the dataset to be
analyzed.  Alternatively, use \link{prepare_dataset} to automatically
read sample attributes from filenames.  If more than one locus is to be
analyzed from a single sequencer sample (i.e., multiplexed samples), either
the \code{locusmap} argument to \code{prepare_dataset} can be used, or \code{load_dataset}
with an explicit mapping of loci to files.
}
\details{
Columns Required:
\itemize{
\item Filename: path to each data file
\item Replicate: identifier for repeated samples; use blanks if not applicable
\item Sample: identifier for a given biological sample
\item Locus: locus identifier matching that used in the locus attributes table
(see \link{load_locus_attrs})
}
}
