% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_seqs}
\alias{load_seqs}
\title{Load vector of sequences from FASTA/FASTQ file}
\usage{
load_seqs(fp)
}
\arguments{
\item{fp}{path to sequence file}
}
\value{
vector of sequences
}
\description{
Load a vector of character sequences from the given path.  This is just a
wrapper around dnar to choose the parser based on filename.  Only the
sequences are returned, not IDs or quality scores.
}
