% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{logmsg}
\alias{logmsg}
\title{Write Log Message}
\usage{
logmsg(msg, col2 = as.character(Sys.time()), end = "\\n")
}
\arguments{
\item{msg}{text to print.}

\item{col2}{extra text to show at right margin; defaults to current time.}

\item{end}{ending to concatenate to message; defaults to newline character.}
}
\description{
Print a log message to the standard error stream.
}
