% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dataset.R
\name{make.dist_mat_known}
\alias{make.dist_mat_known}
\title{Make distance matrix for samples to known genotypes}
\usage{
make.dist_mat_known(results_summary, genotypes.known,
  dist.func = calc.genotype.distance)
}
\arguments{
\item{results_summary}{cross-sample summary data frame}

\item{genotypes.known}{data frame of known genotypes, with one row per
individual per locus, and columns "Name", "Locus", "Allele1Seq",
"Allele2Seq".}

\item{dist_func}{function to calculate inter-sample distances.  Should take
two vectors, one for each genotype, with two values per locus corresponding
to the two alleles.}
}
\value{
matrix of sample-to-individual distance values, with individuals from
  genotypes.known on rows and samples on columns.
}
\description{
Compare the genotype of each sample in a given results summary to all known
genotypes in the supplied table, and create a sample-to-known-genotype
distance matrix.
}
