% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{make_allele_names}
\alias{make_allele_names}
\title{Create Short Allele Names}
\usage{
make_allele_names(seqs, hash_len = cfg("allele_suffix_len"))
}
\arguments{
\item{hash_len}{number of characters of alphanumeric hash to include as a
suffix.  (See \link[openssl:hash]{openssl::md5}.)}

\item{data}{character vector of sequences}
}
\value{
vector of short names
}
\description{
Autogenerate short names for sequences using sequence length and content.
}
