% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dataset.R
\name{make_dist_mat}
\alias{make_dist_mat}
\title{Make distance matrix for set of samples}
\usage{
make_dist_mat(results_summary, dist.func = calc_genotype_distance)
}
\arguments{
\item{results_summary}{cross-sample summary data frame}

\item{dist.func}{function to calculate inter-sample distances.  Should take
two vectors, one for each genotype, with two values per locus corresponding
to the two alleles.}
}
\value{
matrix of cross-sample distance values
}
\description{
Compare the genotype of each sample to each other sample in a given results
summary data frame, and create a between-sample distance matrix.  This will
be a symmetric matrix, unlike what \code{make_dist_mat_known} produces.
}
