% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{make_read_primer_table}
\alias{make_read_primer_table}
\title{Make data frame of primer info for reads}
\usage{
make_read_primer_table(
  seqs,
  locus_attrs,
  max_mismatches = config.defaults$seq_analysis$max_mismatches,
  primer_action = config.defaults$seq_analysis$primer_action,
  primer_action_fwd = config.defaults$seq_analysis$primer_action_fwd,
  primer_action_rev = config.defaults$seq_analysis$primer_action_rev,
  max_mismatches_fwd = config.defaults$seq_analysis$max_mismatches_fwd,
  max_mismatches_rev = config.defaults$seq_analysis$max_mismatches_rev,
  use_reverse_primers = config.defaults$seq_analysis$use_reverse_primers,
  reverse_primer_r1 = config.defaults$seq_analysis$reverse_primer_r1
)
}
\arguments{
\item{seqs}{character vector of read sequences}

\item{locus_attrs}{data frame of locus attributes}

\item{max_mismatches}{integer number of mismatches allowed when checking
primers against reads}

\item{primer_action}{how should reads be modified based on matched primers?
Can be "none", "keep" (trim sequences to the primers but keep the primers
as-is), "replace" (trim sequences \emph{and} replace the matched primer region
with the sequence from \code{locus_attrs}), or "remove" (trim sequences to
exclude the primer region).}

\item{primer_action_fwd}{a \code{primer_action} value for the forward primer}

\item{primer_action_rev}{a \code{primer_action} value for the reverse primer}

\item{max_mismatches_fwd}{a \code{max_mismatches} value for the forward primer}

\item{max_mismatches_rev}{a \code{max_mismatches} value for the reverse primer}

\item{use_reverse_primers}{use reverse primers, or just forward?}

\item{reverse_primer_r1}{is each reverse primer in \code{locus_attrs} given in its
orientation on the forward read?  This is used to determine how the primers
and reads should be reverse complemented before comparing.}
}
\value{
data frame of matched primer information, one row for each input
sequence
}
\description{
\code{make_read_primer_table} take read sequences and per-locus primer
information and produces a data frame of matching information for each read.
Mismatches are allowed, but not indels.  The rows in the output data frame
will correspond exactly to the read vector given.
}
\details{
Output Columns:
\itemize{
\item SeqOrig: Original read sequence
\item FwdStart: start position of forward primer match
\item Fwdstop: end position of forward primer match
\item FwdMismatches: number of mismatches for the forward primer
\item FwdLocus: name of the Locus whose forward primer matched
\item RevStart: start position of reverse primer match
\item RevStop: end position of reverse primer match
\item RevMismatches: number of mismatches in the reverse primer
\item RevLocus: name of the Locus whose reverse primer matched
\item MatchingLocus: name of Locus matching in forward or both (if
\code{use_reverse_primers}) directions
\item Seq: modified sequence based on \code{primer_action} argument(s), if
applicable
}

Optionally reads can be modified based on the matched primer sequences in one
or both directions; see the \code{primer_action} arguments for those options.  All
positions in the returned data frame are indexed from 1 and are oriented in
the same direction as the unmodified reads.

The comparison of reads to primers is not exhaustive.  This method gathers
exact matches first and searches the remaining reads for inexact matches in
order of decreasing abundance of exact matches, taking the first
sufficiently-similar match for each read.  For \code{max_mismatches} values well
below the similarity between any two primer sequences (as will typically be
the case) this should result in the best-matching primer being reported for
each read, but is much faster than doing an exhaustive all-to-all comparison.
}
