% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsat.R
\docType{package}
\name{microsat-package}
\alias{microsat}
\alias{microsat-package}
\title{Analyze Microsatellites}
\description{
Analyze DNA microsatellites in high-throughput sequencing datasets.
}
\author{
\strong{Maintainer}: Jesse Connell \email{ancon@upenn.edu}

}
