% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{name_alleles_in_table}
\alias{name_alleles_in_table}
\title{Name allele sequences in genotype data frame}
\usage{
name_alleles_in_table(data, known_alleles = NULL, name_args = list())
}
\arguments{
\item{data}{data frame containing Allele1Seq and Allele2Seq colums such as
the first list item produced by \code{\link{analyze_dataset}}.}

\item{known_alleles}{data frame of custom allele names as defined for
\code{\link{load_allele_names}}.  if NULL only automatically generated
names will be used.}

\item{name_args}{list of additional arguments to \code{make_allele_name}.}
}
\value{
data frame provided with Allele1Name and Allele2Name columns added
}
\description{
Add Allele1Name and Allele2Name columns matching Allele1Seq and Allele2Seq in
the given data frame.  Names from the given known_alleles data frame will be
used for recognized sequences.
}
