% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{parse_config}
\alias{parse_config}
\title{Parse a configuration data frame into list}
\usage{
parse_config(cfg_table)
}
\arguments{
\item{cfg_table}{data frame of CHIIMP configuration options}
}
\value{
list with one item per row in the input data frame, with each value
parsed according to the function name in the Parser column
}
\description{
\code{parse_config} takes each row of a data frame of configuration options and
defines a list item for it, using the Key column for each name and parsing
each value according to the Parser column.
}
