% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot.heatmap.homozygous}
\alias{plot.heatmap.homozygous}
\title{Plot heatmap of homozygous samples}
\usage{
\method{plot}{heatmap.homozygous}(results_summary, ...)
}
\arguments{
\item{results_summary}{cross-sample summary data frame as produced by
\code{analyze.dataset}.}

\item{...}{additional arguments passed to \code{plot.heatmap}.}
}
\description{
Given a cross-sample summary data frame as produced by
\code{analyze.dataset}, plot a heatmap showing which samples appear
homozygous, with sample identifiers on rows and loci on columns.
}
