% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot.heatmap.prominent_seqs}
\alias{plot.heatmap.prominent_seqs}
\title{Plot heatmap of samples with multiple prominent sequences}
\usage{
\method{plot}{heatmap.prominent_seqs}(results_summary, ...)
}
\arguments{
\item{results_summary}{cross-sample summary data frame as produced by
\code{analyze.dataset}.}

\item{...}{additional arguments passed to \code{plot.heatmap}.}
}
\description{
Given a cross-sample summary data frame as produced by
\code{analyze.dataset}, plot a heatmap showing samples with more than two
prominent sequences in their analysis output, with sample identifiers on rows
and loci on columns.
}
