% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot.heatmap.proportions}
\alias{plot.heatmap.proportions}
\title{Plot heatmap of proportion of allele sequence counts}
\usage{
\method{plot}{heatmap.proportions}(results_summary, ...)
}
\arguments{
\item{results_summary}{cross-sample summary data frame as produced by
\code{analyze.dataset}.}

\item{...}{additional arguments passed to \code{plot.heatmap}.}
}
\description{
Given a cross-sample summary data frame as produced by
\code{analyze.dataset}, plot a heatmap showing the proportion of matching
sequences for the identified alleles versus all matching sequences, with
sample identifiers on rows and loci on columns.
}
