% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot.heatmap}
\alias{plot.heatmap}
\title{Render heatmap of STR attribute across samples and loci}
\usage{
\method{plot}{heatmap}(results_summary, attribute,
  label.by = c("Allele1Length", "Allele2Length"), color = c("white",
  "pink"), breaks = NA, ...)
}
\arguments{
\item{results_summary}{cross-sample summary data frame as produced by
\code{analyze.dataset}.}

\item{attribute}{character name of column in results_summary to use.}

\item{label.by}{vector of column names to use when writing the genotype
summary values on top of the heatmap cells (default: allele lengths).}

\item{color}{vector of colors passed to \code{pheatmap}.}

\item{breaks}{vector of breakpoints passed to \code{pheatmap}.}

\item{...}{additional arguments to \code{pheatmap}.}
}
\description{
Given a cross-sample summary data frame as produced by \code{analyze.dataset}
and the name of a column (e.g., Stutter, Homozygous, ProminentSequences),
plot a heatmap of the values for that attribute, with sample identifiers on
rows and loci on columns.  The attribute will be coerced to numeric.
}
