% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot.heatmap.stutter}
\alias{plot.heatmap.stutter}
\title{Plot heatmap of suspected PCR stutter}
\usage{
\method{plot}{heatmap.stutter}(results_summary, ...)
}
\arguments{
\item{results_summary}{cross-sample summary data frame as produced by
\code{analyze.dataset}.}

\item{...}{additional arguments passed to \code{plot.heatmap}.}
}
\description{
Given a cross-sample summary data frame as produced by
\code{analyze.dataset}, plot a heatmap showing which samples had alleles
ignored due to suspected PCR stutter, with sample identifiers on rows and
loci on columns.
}
