% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_heatmap_artifacts}
\alias{plot_heatmap_artifacts}
\title{Plot heatmap of suspected PCR artifacts}
\usage{
plot_heatmap_artifacts(results, ...)
}
\arguments{
\item{results}{combined results list}

\item{...}{additional arguments passed to \code{\link{plot_heatmap}}.}
}
\description{
Given a cross-sample summary data frame as produced by
\code{\link{analyze_dataset}}, plot a heatmap showing which samples had
alleles ignored due to a suspected PCR artifact, with sample identifiers on
rows and loci on columns.
}
\seealso{
\code{\link{plot_heatmap}}
}
