% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_heatmap_homozygous}
\alias{plot_heatmap_homozygous}
\title{Plot heatmap of homozygous samples}
\usage{
plot_heatmap_homozygous(results, ...)
}
\arguments{
\item{results}{combined results list}

\item{...}{additional arguments passed to \code{\link{plot_heatmap}}.}
}
\description{
Given a cross-sample summary data frame as produced by
\code{\link{analyze_dataset}}, plot a heatmap showing which samples appear
homozygous, with sample identifiers on rows and loci on columns.
}
\seealso{
\code{\link{plot_heatmap}}
}
