% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_heatmap_proportions}
\alias{plot_heatmap_proportions}
\title{Plot heatmap of proportion of allele sequence counts}
\usage{
plot_heatmap_proportions(results, ...)
}
\arguments{
\item{results}{combined results list}

\item{...}{additional arguments passed to \code{\link{plot_heatmap}}.}
}
\description{
Given a cross-sample summary data frame as produced by
\code{\link{analyze_dataset}}, plot a heatmap showing the proportion of
matching sequences for the identified alleles versus all matching sequences,
with sample identifiers on rows and loci on columns.
}
\seealso{
\code{\link{plot_heatmap}}
}
