% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{prepare.dataset}
\alias{prepare.dataset}
\title{Load table of locus attributes}
\usage{
prepare.dataset(directory, pattern, ord = c(1, 2, 3))
}
\arguments{
\item{directory}{location to search for matching data files.}

\item{pattern}{regular expression to use for parsing filenames.}

\item{ord}{integer vector giving order of fields (replicate/sample/locus) in
filenames.}
}
\value{
data frame of metadata for all files found
}
\description{
Load a tab-separated table of locus attributes to use for analysis.
TODO rewrite this mess!
}
