% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{Load table of locus attributes}
\usage{
prepare_dataset(dp, pattern, ord = c(1, 2, 3))
}
\arguments{
\item{dp}{directory path to search for matching data files.}

\item{pattern}{regular expression to use for parsing filenames.}

\item{ord}{integer vector giving order of the fields Replicate, Sample, and
Locus in filenames.  For example, if Locus is the first field followed by
Replicate and Sample, set \code{ord=c(3, 1, 2)}.}
}
\value{
data frame of metadata for all files found
}
\description{
Load a tab-separated table of locus attributes to use for analysis.
}
