% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\docType{data}
\name{RAW_NT}
\alias{RAW_NT}
\title{IUPAC DNA nucleotides as raw bytes}
\format{
An object of class \code{raw} of length 17.
}
\usage{
RAW_NT
}
\description{
\code{RAW_NT} is a named vector of raw bytes allowing fast comparison of IUPAC
DNA nucleotide characters.  A, C, G, and T are assigned single individual
bits, and the standard IUPAC characters are assigned bitwise combinations of
these.  This way \code{RAW_NT["A"] & RAW_NT["G"]} is zero, while
\code{RAW_NT["A"] & RAW_NT["R"]} is nonzero.  The gap characters - and . are
assigned one remaining bit, leaving 3 bits unspecified.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{     - TGCA
A 0000 0001  01
C 0000 0010  02
G 0000 0100  04
T 0000 1000  08

R 0000 0101  05
Y 0000 1010  0A
S 0000 0110  06
W 0000 1001  09
K 0000 1100  0C
M 0000 0011  03

B 0000 1110  0E
D 0000 1101  0D
H 0000 1011  0B
V 0000 0111  07

N 0000 1111  0F

- 0001 0000  10
. 0001 0000  10
}\if{html}{\out{</div>}}
}
\keyword{datasets}
