% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{remove_shared_root_dir}
\alias{remove_shared_root_dir}
\title{Remove shared path from file paths}
\usage{
remove_shared_root_dir(fps_full)
}
\arguments{
\item{fps_full}{character vector of file paths.}
}
\value{
character vector of same length as input, with any common directory
  structure trimmed off.
}
\description{
For the given character vector of file paths, create a modified version with
any common prefix path removed.
}
