% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsat.R
\name{render_report}
\alias{render_report}
\title{Render Microsatellite Report Document}
\usage{
render_report(results, config)
}
\arguments{
\item{results}{list of microsatellite analysis results as produced by
\code{\link{full_analysis}}.}

\item{config}{list of configuration options (see
\code{\link{config.defaults}}).}
}
\description{
Write an HTML report summarizing the results of a full microsatellite
analysis.
}
