% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_idents}
\alias{report_idents}
\title{Create identification summary table}
\usage{
report_idents(results, closest, allele.names = NULL, hash.len = 6,
  na.replicates = "")
}
\arguments{
\item{results}{results list as produced by \code{\link{summarize_dataset}}.}

\item{closest}{list of closest matches as produced by
\code{\link{find_closest_matches}}.}

\item{allele.names}{optional data frame of custom allele names.  If present
the columns "Seq" and "Name" will be used to map exact sequences to custom
names.}

\item{hash.len}{number of characters of the sequence has to include when
auto-generating allele names.  If set to zero the hash suffix is not
included.}

\item{na.replicates}{text to replace NA entries with for the Replicates
column.}
}
\value{
data frame showing summary of sample genotypes with interleved
  genotypes for similar known individuals.
}
\description{
Report the genotypes present in a processed dataset, paired with close
matches to known individuals, converting sequences to short names with
optional custom naming via \code{\link{report_genotypes}}.
}
