% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{revcmp}
\alias{revcmp}
\title{Reverse complement sequences}
\usage{
revcmp(txt)
}
\arguments{
\item{txt}{character vector of sequences}
}
\value{
character vector of reverse complements
}
\description{
Each entry in the input character vector is reversed and nucleotide
characters replaced with their complements (leaving any other text characters
unchanged).
}
