% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sample.R
\docType{data}
\name{sample_analysis_funcs}
\alias{sample_analysis_funcs}
\title{Versions of analyze_sample}
\format{
An object of class \code{character} of length 3.
}
\usage{
sample_analysis_funcs
}
\description{
This list shows the different versions of \code{\link{analyze_sample}}
recognized for use via configuration in full_analysis.  See also
\code{\link{sample_summary_funcs}} and \code{\link{config.defaults}}.
}
\keyword{datasets}
