% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save.allele_seqs}
\alias{save.allele_seqs}
\title{Save identified alleles to FASTA files}
\usage{
save.allele_seqs(results_summary, dp)
}
\arguments{
\item{results_summary}{summary data frame as produced by
\code{analyze.dataset}.}

\item{dp}{output directory path to use for all files.}
}
\description{
Take the alleles identified by \code{analyze.dataset} in the summary data
frame and save each entry to a separate FASTA file.  Samples identified as
homozygous will have one sequence written rather than two.  Entries with no
identified alleles will be skipped.
}
