% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save.dist_mat}
\alias{save.dist_mat}
\title{Save sample distance matrix to text file}
\usage{
save.dist_mat(dist_mat, fp)
}
\arguments{
\item{dist_mat}{matrix produced by \code{summarize.dataset}.}

\item{fp}{output file path.}
}
\description{
Save the inter-sample distance matrix produced by \code{summarize.dataset} to
the specified file path in CSV format.
}
