% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save.results_summary}
\alias{save.results_summary}
\title{Save dataset summary to text file}
\usage{
save.results_summary(results_summary, fp)
}
\arguments{
\item{results_summary}{summary data frame as produced by
\code{analyze.dataset}.}

\item{fp}{output file path.}
}
\description{
Save the dataset summary produced by \code{analyze.dataset} to the specified
file path in CSV format.
}
