% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_alignments}
\alias{save_alignments}
\title{Save alignments to FASTA files}
\usage{
save_alignments(alignments, dp)
}
\arguments{
\item{alignments}{list of MSA alignment objects, such as created by
\link{summarize_dataset} via \link{align_alleles}.  The name of each alignment will
be used for its filename.}

\item{dp}{output directory path.}
}
\description{
Take a list of alignments, one per locus, and save each to a separate FASTA
file in a specified directory.  If any of the per-locus alignment objects is
NA it will be skipped.  These are produced by \link{summarize_dataset} via
\link{align_alleles}.
}
