% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiimp.R
\name{save_data}
\alias{save_data}
\title{Save Microsatellite Analysis to Disk}
\usage{
save_data(results, config)
}
\arguments{
\item{results}{list of microsatellite analysis results as produced by
\code{\link{full_analysis}}.}

\item{config}{list of configuration options (see
\code{\link{config.defaults}}).}
}
\description{
Save the various parts of a finished microsatellite analysis to data files
and plot images.
}
