% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_dataset}
\alias{save_dataset}
\title{Save table of sample attributes}
\usage{
save_dataset(data, fp, ...)
}
\arguments{
\item{data}{data frame of sample attributes as produced by
\code{\link{prepare_dataset}} or \code{\link{load_dataset}}.}

\item{fp}{path to text file.}

\item{...}{additional arguments passed to \code{\link[utils]{read.table}}.}
}
\description{
Save a comma-separated table of sample attributes.
}
