% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_dist_mat}
\alias{save_dist_mat}
\title{Save sample distance matrix to text file}
\usage{
save_dist_mat(dist_mat, fp)
}
\arguments{
\item{dist_mat}{matrix produced by \code{\link{summarize_dataset}} via
\code{\link{make_dist_mat}}.}

\item{fp}{output file path.}
}
\description{
Save the inter-sample distance matrix produced by
\code{\link{summarize_dataset}} to the specified file path in CSV format.
}
