% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_histograms}
\alias{save_histograms}
\title{Save sequence histogram visualizations to image files}
\usage{
save_histograms(results, dp, image.func = "png", width = 1600,
  height = 1200, res = 150)
}
\arguments{
\item{results}{list of results as created by \code{analyze.dataset}.}

\item{dp}{output directory path.}

\item{image.func}{name of function to call for saving each image.}

\item{width}{integer width of image.}

\item{height}{integer height of image.}

\item{res}{integer resolution of image in PPI.}
}
\description{
Take a full results list and save a histogram of each sample to a separate
image file in a specified directory.
}
