% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_results_summary}
\alias{save_results_summary}
\title{Save dataset summary to text file}
\usage{
save_results_summary(results_summary, fp)
}
\arguments{
\item{results_summary}{summary data frame as produced by \link{analyze_dataset}.}

\item{fp}{output file path.}
}
\description{
Save the dataset summary produced by \link{analyze_dataset} to the specified file
path in CSV format.
}
