% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_sample_data}
\alias{save_sample_data}
\title{Save per-sample processed data to text files}
\usage{
save_sample_data(results_data, dp)
}
\arguments{
\item{results_data}{list of per-sample data frames as produced by
\code{\link{analyze_dataset}}.}

\item{dp}{output directory path to use for all files.}
}
\description{
Save each per-sample data frame produced by \code{\link{analyze_dataset}} to
a separate file in the specified directory path, in CSV format.
}
