% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiimp.R
\name{setup_package}
\alias{setup_package}
\title{Set up CHIIMP package environment}
\usage{
setup_package()
}
\description{
This loads \link{CFG_DEFAULTS} from a CSV file when the chiimp package is loaded,
and applies the configuration options globally.
}
