% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{str_hist_render}
\alias{str_hist_render}
\title{Draw prepared histogram data}
\usage{
str_hist_render(bars, main, xlim, cutoff_fraction)
}
\description{
Render prepared histogram data to the display device.
}
