% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{str_hist_setup}
\alias{str_hist_setup}
\title{Prepare histogram data}
\usage{
str_hist_setup(seq_data, sample_data = NULL)
}
\description{
Create the counts-by-length data frames for \code{\link{histogram}}.
}
