% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{str_hist_setup_legend}
\alias{str_hist_setup_legend}
\title{Setup display attributes for STR histogram}
\usage{
str_hist_setup_legend(bars)
}
\description{
Create data frame of plot attributes to use in STR histogram.
}
