% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dataset.R
\name{summarize.dataset}
\alias{summarize.dataset}
\title{Add further summaries to analyzed dataset}
\usage{
summarize.dataset(results, genotypes.known = NULL)
}
\arguments{
\item{results}{list containing summary data frame and sample-specific data
frames as produced by \code{analyze.dataset}.}

\item{genotypes.known}{optional data frame of known genotypes that should be
compared to the observed genotypes in the results.  If provided
dist_mat_known will be present in the output.}
}
\value{
expanded list with additional summaries.
}
\description{
Take a results list as produced by \code{analyze.dataset} and add additional
entries for inter-sample and inter-locus analyses:
}
\details{
Additional entries in the returned list:
\itemize{
\item alignments: inter-allele alignments for each locus
\item dist_mat: inter-sample distance matrix
\item dist_mat_known: if genotypes.known is given, this distance matrix of
sample-to-individual values will be present.
}
}
