% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dataset.R
\name{summarize_attribute}
\alias{summarize_attribute}
\title{Create Summary Attribute Table}
\usage{
summarize_attribute(results_summary, attrib, repeats = 2)
}
\arguments{
\item{results_summary}{cross-sample summary data frame as produced by
\code{\link{analyze_dataset}}.}

\item{attrib}{name of column from \code{results_summary} to tabulate.}

\item{repeats}{number of times to repeat each column, for matching with two
alleles per entry.}
}
\value{
data frame of attribute across samples and loci.
}
\description{
Tabulate a single arbitrary attribute across loci, assuming repeats by two
for the alleles.  This is used for color-coding summary heatmaps (see
\code{\link{plot_heatmap}}) on top of the attribute values, like
\code{Homozygous} or \code{ProminentSeqs}.
}
