% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dataset.R
\name{summarize_genotypes}
\alias{summarize_genotypes}
\title{Create Summary Genotype Table}
\usage{
summarize_genotypes(results_summary, vars = c("Allele1Seq",
  "Allele2Seq"))
}
\arguments{
\item{results_summary}{cross-sample summary data frame as produced by
\code{\link{analyze_dataset}}.}

\item{vars}{vector of column names in \code{results_summary} to use for the
summary.  Defaults to the sequence content for the two alleles.}
}
\value{
data frame of genotypes across samples and loci.
}
\description{
Tabulate genotypes across an analyzed dataset, with samples on rows and loci
on columns.  Which attributes will actually be used for the table is
configurable.
}
