% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dataset.R
\name{summarize_genotypes_known}
\alias{summarize_genotypes_known}
\title{Create Summary Genotype Table for Known Genotypes}
\usage{
summarize_genotypes_known(genotypes_known, tbl_genotypes = NULL)
}
\arguments{
\item{genotypes_known}{table of known individuals' genotypes as loaded via
\link{load_genotypes}.}

\item{tbl_genotypes}{existing genotype summary table to use for locus
selection and column ordering.}
}
\value{
data frame of genotypes across individuals and loci.
}
\description{
Tabulate genotypes across an analyzed dataset, with samples on rows and loci
on columns.  Which attributes will actually be used for the table is
configurable.
}
