% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{tabulate_allele_names}
\alias{tabulate_allele_names}
\title{Wide table of allele names vs loci}
\usage{
tabulate_allele_names(data, extra_cols = NULL)
}
\arguments{
\item{data}{data frame containing Allele1Name and Allele2Name columns such as
the first list item produced by \code{\link{analyze_dataset}}.  If allele
names are not yet present call \code{\link{name_alleles_in_table}}.}

\item{extra_cols}{names or index values of additional columns from input data
frame to be kept in output data frame.  These should be consistent across
loci for a given entry.}
}
\value{
wide format data frame with sample entries on rows and loci on
  columns.  An ID column will label sample entries by whichever columns were
  provided in the input (see \code{\link{make_entry_id}}).  Empty strings are
  given for NA allele names, while NA is given for any implicitly missing
  locus/sample combinations.
}
\description{
Allele pairs are shown in a standardized order with homozygous entries shown
twice.  NA allele names in the input are converted to empty strings while NA
is given for missing sample/locus combinations.
}
