% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dataset.R
\name{tally_cts_per_locus}
\alias{tally_cts_per_locus}
\title{Create Counts-per-Locus Table}
\usage{
tally_cts_per_locus(results)
}
\arguments{
\item{results}{list containing summary data frame and sample-specific data
frames as produced by \code{\link{analyze_dataset}}.}
}
\value{
Data frame of read counts with samples on rows and loci on columns.
  Addition columns at the left specify the number of reads matching any known
  locus (\code{Total}) and the number of reads matching the expected locus
  (\code{Matching}).
}
\description{
Produce a table of read counts matching (by primer) each locus per sample.
}
