% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_helper_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Helper Data for Tests}
\format{An object of class \code{list} of length 18.}
\usage{
test_data
}
\description{
This list is a bundle of shared data and functions for running unit tests.
}
\keyword{datasets}
