% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{version_compare}
\alias{version_compare}
\title{Compare two version strings}
\usage{
version_compare(ver1txt, ver2txt)
}
\arguments{
\item{ver1txt}{single character string for a software version}

\item{ver2txt}{single character string for a software version}
}
\value{
a single character representing if if the first version is later
(\code{>}), earlier (\code{<}), or equivalent (\code{=}) to the second version
}
\description{
\code{version_compare} takes two strings for software versions(e.g. "1.0.2") and
returns \code{>}, \code{<}, or \code{=} depending on whether the first version is later,
earlier, or equivalent to the second version.  For example, "1.2.0" >
"1.1.5".
}
\details{
If there are fewer fields in one string than the other, the one with fewer is
padded with zeros for the least-significant positions.  For example 1.2.3 vs
1.2 is equivalent to 1.2.3 vs 1.2.0.
}
