% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_dataset.R
\name{analyze_dataset}
\alias{analyze_dataset}
\title{Analyze all samples in a dataset}
\usage{
analyze_dataset(dataset, locus_attrs, nrepeats, ncores = 0, summary_args,
  summary.function = summarize_sample, known_alleles = NULL)
}
\arguments{
\item{dataset}{data frame of sample details as produced by
\code{\link{prepare_dataset}}.}

\item{locus_attrs}{data frame of locus attributes as produced by
\code{\link{load_locus_attrs}}.}

\item{nrepeats}{number of repeats of each locus' motif to require for a
match (see \code{\link{analyze_sample}}).}

\item{ncores}{integer number of CPU cores to use in parallel for sample
analysis.  Defaults to one less than half the number of detected cores with
a minimum of 1.  If 1, the function will run without using the
\code{parallel} package.}

\item{summary_args}{list of supplemental arguments to
\code{summary.function}.}

\item{summary.function}{function to use when summarizing each sample's full
details into the standard attributes.  Defaults to
\code{\link{summarize_sample}}.}

\item{known_alleles}{data frame of custom allele names as defined for
\code{\link{load_allele_names}}.  if NULL only the names automatically
generated for the dataset summary will be used.}
}
\value{
list of results, with \code{summary} set to the single summary data
  frame and \code{data} the per-sample data frames.
}
\description{
Load all samples for a given dataset and produce a list of processed samples
and a summary data frame showing each sample's summary row-by-row.  The
entries in the processed-samples list and the rows in the summary data frame
will be sorted according to the ordering of loci in \code{locus_attrs} and
by the sample attributes.
}
