% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_dataset.R
\name{analyze_dataset}
\alias{analyze_dataset}
\title{Analyze all samples in a dataset}
\usage{
analyze_dataset(
  dataset,
  locus_attrs,
  nrepeats = cfg("min_motif_repeats"),
  stutter.count.ratio_max = cfg("max_stutter_ratio"),
  artifact.count.ratio_max = cfg("max_artifact_ratio"),
  use_reverse_primers = cfg("use_reverse_primers"),
  reverse_primer_r1 = cfg("reverse_primer_r1"),
  max_mismatches = cfg("max_mismatches"),
  primer_action = cfg("primer_action"),
  max_mismatches_fwd = cfg("max_mismatches_fwd"),
  max_mismatches_rev = cfg("max_mismatches_rev"),
  primer_action_fwd = cfg("primer_action_fwd"),
  primer_action_rev = cfg("primer_action_rev"),
  ncores = cfg("ncores"),
  analysis_opts,
  summary_opts,
  analysis_function = analyze_sample,
  summary_function = summarize_sample,
  known_alleles = NULL,
  name_args = list()
)
}
\arguments{
\item{dataset}{data frame of sample details as produced by
\code{\link{prepare_dataset}}.}

\item{locus_attrs}{data frame of locus attributes as produced by
\code{\link{load_locus_attrs}}.}

\item{nrepeats}{number of repeats of each locus' motif to require for a
match (see \code{\link{analyze_seqs}}).}

\item{stutter.count.ratio_max}{highest ratio of read counts for second most
frequent sequence to the most frequent where the second will be
considered stutter (see \code{\link{analyze_seqs}}).}

\item{artifact.count.ratio_max}{as for \code{stutter.count.ratio_max} but for
non-stutter artifact sequences.}

\item{use_reverse_primers}{consider the ReversePrimer column from the locus
attributes for locus-matching?}

\item{reverse_primer_r1}{Is each reverse primer given in its orientation on
the forward read?  (See \code{\link{analyze_seqs}})}

\item{ncores}{integer number of CPU cores to use in parallel for sample
analysis.  Defaults to one less than half the number of detected cores with
a minimum of 1.  If 1, the function will run without using the
\code{parallel} package.}

\item{analysis_opts}{list of supplemental arguments to
\code{analysis_function}.}

\item{summary_opts}{list of supplemental arguments to
\code{summary_function}.}

\item{analysis_function}{function to use when analyzing each sample's data
frame into the filtered version  Defaults to \code{\link{analyze_sample}}.}

\item{summary_function}{function to use when summarizing each sample's full
details into the standard attributes.  Defaults to
\code{\link{summarize_sample}}.}

\item{known_alleles}{data frame of custom allele names as defined for
\code{\link{load_allele_names}}.  if NULL only the names automatically
generated for the dataset summary will be used.}

\item{name_args}{list of additional arguments to
\code{\link{make_allele_name}}.}
}
\value{
list of results, with \code{summary} set to the single summary data
  frame, \code{files} the processed sequence files, and \code{samples} the
  per-sample data frames.
}
\description{
Load all samples for a given dataset and produce a summary data frame showing
each sample's summary row-by-row, a list of processed input files, and a list
of processed samples.  The entries in the processed-samples list and the rows
in the summary data frame will be sorted according to the ordering of loci in
\code{locus_attrs} and by the sample attributes.  Processed files are stored
separately (as there may be multiple samples per file) and named by input
file path.  An error is thrown if any locus entries in the given dataset are
not found in the locus attributes data frame.
}
