% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sample.R
\name{analyze_sample}
\alias{analyze_sample}
\title{Analyze a set of STR sequences}
\usage{
analyze_sample(seqs, locus_attrs, nrepeats)
}
\arguments{
\item{seqs}{character vector containing sequences.}

\item{locus_attrs}{data frame of attributes for loci to look for.}

\item{nrepeats}{number of repeats of each locus" motif to require for a
match.}
}
\value{
data frame of dereplicated sequences with added annotations.
}
\description{
Dereplicates the given sequences and annotates any STR sequences found,
returning the processed sample as a data frame with one row per unique
sequence, sorted by count.  At this stage no information is filtered out, and
all loci are treated equally.
}
\details{
Columns in the returned data frame:
\itemize{
\item Seq: sequence text for each unique sequence
\item Count: integer count of occurrences of this exact sequence
\item Length: integer sequence length
\item MatchingLocus: factor for the name of the locus matching each sequence,
by checking the primer
\item MotifMatch: logical -- are there are least \code{nrepeats} perfect
adjacent repeats of the STR motif for the matching locus?
\item LengthMatch: logical -- is the sequence length within the expected range
for the matching locus?
\item Stutter: integer: for any sequence that looks like potential PCR stutter,
the index of the row that may be the source of the stutter band.
\item FractionOfTotal: numeric fraction of the number of sequences
represented by each unique sequence compared to the total.
\item FractionOfLocus: numeric fraction of the number of sequences represented
by each unique sequence compared to the total for that particular
matching locus.
}
}
