% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sample.R
\name{analyze_sample}
\alias{analyze_sample}
\title{Analyze sequence table and categorize sequences}
\usage{
analyze_sample(sample.data, sample.attrs, fraction.min)
}
\arguments{
\item{sample.data}{data frame of processed data for sample as produced by
\code{\link{analyze_seqs}}.}

\item{sample.attrs}{list of sample attributes, such as the rows produced by
\code{\link{prepare_dataset}}.  Used to select the locus name to filter on.}

\item{fraction.min}{numeric threshold for the minimum proportion of counts a
given entry must have, compared to the total matching all criteria for that
locus, to be considered as a potential allele.}
}
\value{
filtered version of \code{sample.data} with added Category column.
}
\description{
Converts a full STR sequence data frame into a per-locus version and adds a
Category factor column to designate which sequences look like alleles,
artifacts, etc.  At this stage the summary is prepared for a single specific
locus, in contrast to \code{\link{analyze_seqs}}.
}
\details{
Factor levels in the added Category column:
\itemize{
\item Allele: An identified allele sequence.  There will be between zero and two
of these.
\item Prominent: Any additional sequences beyond two called alleles that match
all requirements (sequences that match all locus attributes, do not appear
artifactual, and are above a given fraction of filtered reads).
\item Insignificant: Sequences with counts below the \code{fraction.min}
threshold.
\item Ambiguous: Sequences passing the \code{fraction.min} threshold but with
non-ACTG characters such as N, as defined by the Ambiguous column.
\item Stutter: Sequences passing the \code{fraction.min} threshold but matching
stutter sequence criteria as defined by the Stutter column.
\item Artifact: Sequences passing the \code{fraction.min} threshold but matching
non-stutter artifact sequence criteria as defined by the Artifact column.
}
}
