% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_seqs.R
\name{analyze_seqs}
\alias{analyze_seqs}
\title{Analyze a set of STR sequences}
\usage{
analyze_seqs(
  seqs,
  locus_attrs,
  nrepeats = cfg("min_motif_repeats"),
  max_stutter_ratio = cfg("max_stutter_ratio"),
  artifact.count.ratio_max = cfg("max_artifact_ratio"),
  ...
)
}
\arguments{
\item{seqs}{character vector containing sequences.}

\item{locus_attrs}{data frame of attributes for loci to look for.}

\item{nrepeats}{number of repeats of each locus' motif to require for a
match.}

\item{max_stutter_ratio}{highest ratio of read counts for second most
frequent sequence to the most frequent where the second will be
considered stutter.}

\item{artifact.count.ratio_max}{as for \code{max_stutter_ratio} but for
non-stutter artifact sequences.}

\item{...}{additional arguments for \link{make_read_primer_table}}
}
\value{
data frame of dereplicated sequences with added annotations.
}
\description{
Dereplicates the given sequences and annotates any STR sequences found,
returning the processed data as a data frame with one row per unique
sequence, sorted by count.  At this stage no information is filtered out, and
all loci are treated equally.
}
\details{
Columns in the returned data frame:
\itemize{
\item \code{Seq}: sequence text for each unique sequence
\item \code{Count}: integer count of occurrences of this exact sequence
\item \code{Length}: integer sequence length
\item \code{MatchingLocus}: factor for the name of the locus matching each
sequence, by checking the primer
\item \code{MotifMatch}: logical: are there are least \code{nrepeats} perfect
adjacent repeats of the STR motif for the matching locus?
\item \code{LengthMatch}: logical: is the sequence length within the expected
range for the matching locus?
\item \code{Ambiguous}: logical: are there unexpected characters in the sequence
content?
\item \code{Stutter}: integer: for any sequence that looks like potential PCR
stutter, the index of the row that may be the source of the stutter band.
\item \code{Artifact}: integer: for any sequence that looks like potential PCR
artifact (other than stutter), the index of the row that may be the source
of the stutter band.
\item \code{FractionOfTotal}: numeric fraction of the number of sequences
represented by each unique sequence compared to the total.
\item \code{FractionOfLocus}: numeric fraction of the number of sequences
represented by each unique sequence compared to the total for that
particular matching locus.
}
}
\examples{
# Starting from non-locus-specific sequences,
# a locus attributes table, and requiring
# three side-by-side motif repeats to register
# as a motif match for a locus,
raw_seq_vector <- c(test_data$seqs1$A, test_data$seqs1$B)
locus_attrs <- test_data$locus_attrs
num_adjacent_repeats <- 3
# Convert the character vector of sequences
# into a data frame with one row per
# unique sequence.
seq_data <- analyze_seqs(raw_seq_vector,
                         locus_attrs,
                         num_adjacent_repeats)
}
