% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiimp.R
\docType{package}
\name{chiimp-package}
\alias{chiimp}
\alias{chiimp-package}
\title{Analyze Microsatellites}
\description{
Analyze DNA microsatellites in high-throughput sequencing datasets.
}
\details{
Starting from file inputs and producing file outputs, the overall workflow
(handled by \code{\link{full_analysis}} as a configuration-driven wrapper for
the entire process) is:
\itemize{
\item Load input data.  The input spreadsheets are text files using
comma-separated values (CSV).
\itemize{
\item Load data frame of sample information from a spreadsheet via
\code{\link{load_dataset}} or directly from filenames via
\code{\link{prepare_dataset}}.
\item Load data frame of locus attributes via \code{\link{load_locus_attrs}}
\item Optionally, load data frame of names for allele sequences via
\code{\link{load_allele_names}}.
\item Optionally, load data frame of known genotypes for named individuals via
\code{\link{load_genotypes}}.
}
\item Analyze dataset via \code{\link{analyze_dataset}}
\itemize{
\item Load each sequence data file into a character vector with
\code{\link{load_seqs}} and process into a dereplicated data frame with
\code{\link{analyze_seqs}}.
\item For each sample, filter the sequences from the relevant per-file data
frame to just those matching the expected locus and identify possible
alleles, via \code{\link{analyze_sample}}.  (There may be a many-to-one
relationship of samples to files, for example with sequencer
multiplexing.)
\item Process each per-sample data frames into a summary list of attributes
giving alleles identified and related information, via
\code{\link{summarize_sample}}.
\item Organize \code{analyze_dataset} results into a list of per-file data
frames, a list of per-sample data frames, and a single summary data
frame across all samples.
}
\item Summarize results and add additional comparisons (cross-sample and to
known-individual) via \code{\link{summarize_dataset}}.
\itemize{
\item Tabulate sequence counts per sample matching each locus' primer via
\code{\link{tally_cts_per_locus}}.
\item Align identified alleles for each locus via
\code{\link{align_alleles}}.
\item Create a sample-to-sample distance matrix of allele mismatches via
\code{\link{make_dist_mat}}.
\item If genotypes for known individuals were provided, create a
sample-to-known-individual distance matrix via
\code{\link{make_dist_mat_known}}.
\item If identities of samples were provided, score genotyping success via
\code{\link{match_known_genotypes}} and
\code{\link{categorize_genotype_results}}.
}
\item Save analysis results to files.  Spreadsheets are in CSV format for output
as well as input.  Some output files are in FASTA format (alignments and
alleles) or are PNG images (alignment visualization and sequence count
histograms).  If specified in the configuration, \code{\link{saveRDS}} is
called on the entire output as well, saving to \code{results.rds} by
default.
\item Create an HTML report document summarizing all results.
}

For defaults used in the configuration, see \code{\link{config.defaults}}.

The workflow above outlines CHIIMP's behavior when called as a standalone
program, where \code{\link{main}} loads a configuration file into a nested
list of options and calls \code{\link{full_analysis}}.  The public functions
linked above can also be used idependently; see the documentation and code
examples for the individual functions for more information.
}
\author{
\strong{Maintainer}: Jesse Connell \email{ancon@upenn.edu}

}
