% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\docType{data}
\name{config.defaults}
\alias{config.defaults}
\title{Default microsatellite configuration}
\format{An object of class \code{list} of length 23.}
\usage{
config.defaults
}
\description{
The entries in this list show the build-time configuration defaults for all
aspects of the microsatellite analysis.  These can be overridden by passing a
list to \code{\link{full_analysis}} with entries of the same names, or via
the configuration file passed to \code{\link{main}} and read via
\code{\link{load_config}}.  Check the contents of \code{config.defaults}
itself to see all of the build-time defaults.
}
\details{
Notable Options:
\itemize{
\item dataset_opts:
\itemize{
\item dp: directory path to input sequence files
\item pattern: regular expression for the input filename pattern
\item ord: order of fields in the input filename pattern
}
\item output:
\itemize{
\item dp: directory path for saving output data
}
\item fp_dataset: file path to table of sample attributes to use, rather than
detecting via dataset_opts
\item fp_locus_attrs: file path to locus attributes CSV file
\item fp_genotypes_known: file path to known genotypes CSV file
}
}
\keyword{datasets}
