% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsat.R
\docType{data}
\name{config.defaults}
\alias{config.defaults}
\title{Analyze Microsatellites}
\format{An object of class \code{list} of length 24.}
\usage{
config.defaults
}
\description{
Analyze DNA microsatellites in high-throughput sequencing datasets.
Default microsatellite configuration
}
\details{
The entries in this list show the build-time configuration defaults for all
aspects of the microsatellite analysis.  These can be overridden by passing a
list to \code{full_analysis} with entries of the same names.

Notable Options:
\itemize{
\item dp.data: directory path to input sequence files
\item pattern: regular expression for the input filename pattern
\item ord: order of fields in the input filename pattern
\item dp.output: directory path for saving output data
\item fp.locus_attrs: file path to locus attributes TSV file
\item fp.genotypes.known: file path to known genotypes TSV file
}
}
\keyword{datasets}
