% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_sample.R
\name{find_matching_primer}
\alias{find_matching_primer}
\title{Label STR sample rows by locus}
\usage{
find_matching_primer(sample.data, locus_attrs)
}
\arguments{
\item{sample.data}{data frame of processed sample data.}

\item{locus_attrs}{data frame of attributes for loci to look for.}
}
\value{
factor of locus names corresponding the matched primer sequences.
}
\description{
Return a factor giving the locus name matching each sample data entry"s
matching primer (first found).
}
