% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsat.R
\name{full_analysis}
\alias{full_analysis}
\title{Perform a full microsatellite analysis}
\usage{
full_analysis(config)
}
\arguments{
\item{config}{list of configuration options.  See the summary for
\code{config.defaults} for more details.}
}
\value{
list of results, with the full configuration list included as
  "config."
}
\description{
Given a list of configuration options, run all aspects of a microsatellite
analysis, and save the corresponding output files.
}
