% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiimp.R
\name{full_analysis}
\alias{full_analysis}
\title{Perform a full microsatellite analysis}
\usage{
full_analysis()
}
\value{
list of results, with the full configuration list included as
"config."
}
\description{
Run all aspects of a microsatellite analysis, using the currently-defined
global options (see \link{load_config} and \link{apply_config}), and save the
corresponding output files.
}
\examples{
\dontrun{
# Set up a temporary copy of the CHIIMP test data
example_dir <- tempfile()
dir.create(example_dir)
setwd(example_dir)
test_data$write_seqs(test_data$seqs,
                     "str-dataset",
                      "Replicate1-Sample\%s-\%s.fasta")
locus_attrs_path <- system.file("example_locus_attrs.csv",
                                package = "chiimp")
file.copy(locus_attrs_path, "locus_attrs.csv")
# Run the example analysis
config_path <- system.file("example_config.csv", package = "chiimp")
config <- load_config(config_path)
apply_config(config)
results <- full_analysis()
}

}
