% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{histogram}
\alias{histogram}
\title{Plot histogram of STR sample}
\usage{
histogram(samp, stacked = TRUE, main = NULL, locus.name = NULL,
  sample.summary = NULL)
}
\arguments{
\item{samp}{data frame of dereplicated sequences.}

\item{stacked}{should bars be stacked together to add up to the total count
for each length?  Otherwise they will be overlaid, so each bar height only
corresponds to the count for that exact sequence.}

\item{main}{title of the plot.}

\item{locus.name}{name of the locus to match alleles for.  If unspecified the
locus with the highest matched counts will be used.}

\item{sample.summary}{summary data frame as prepared by
\code{summarize_sample}}
}
\description{
Given a processed STR sample, plot a histogram of counts per sequence length.
}
