% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{histogram}
\alias{histogram}
\title{Plot basic histogram of STR sample}
\usage{
histogram(samp, main = "Number of Reads by Sequence Length",
  locus.name = NULL, sample.summary = NULL, cutoff_fraction = 0.05,
  xlim = range(samp$Length))
}
\arguments{
\item{samp}{data frame of dereplicated sequences.}

\item{main}{title of the plot.}

\item{locus.name}{name of the locus to match alleles for.  If unspecified the
locus with the highest matched counts will be used.}

\item{sample.summary}{summary data frame as prepared by
\code{summarize_sample}}
}
\description{
Given a processed STR sample, plot a histogram of counts per sequence length.
}
