% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_config}
\alias{load_config}
\title{Load configuration file}
\usage{
load_config(fp.config)
}
\arguments{
\item{fp.config}{path to configuration file.}
}
\value{
list of configuration options
}
\description{
Load a YAML-formatted text file of configuration options for microsatellite
analysis.  This is currently just a wrapper around \code{yaml.load()}.
}
