% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_locus_attrs}
\alias{load_locus_attrs}
\title{Load table of locus attributes}
\usage{
load_locus_attrs(fp_locus_attrs, ...)
}
\arguments{
\item{fp_locus_attrs}{path to text file.}

\item{...}{additional arguments passed to \link{load_csv}.}
}
\value{
data frame of locus attributes
}
\description{
Load a comma-separated table of locus attributes to use for analysis.  This
is called automatically during \link{full_analysis}, with the data frame then used
by \link{analyze_seqs} within \link{analyze_dataset}.
}
\details{
Columns Required:
\itemize{
\item \code{Locus}: Unique identifier for a given locus
\item \code{LengthMin}: Minimum known allele sequence length for this locus
\item \code{LengthMax}: Minimum known allele sequence length for this locus
\item \code{LengthBuffer}: Additional sequence length below \code{LengthMin} and above
\code{LengthMax} to accept for a candidate allele
\item \code{Primer}: The forward PCR primer sequence for a given locus, used when
matching sequences to loci
\item \code{ReversePrimer}: The reverse PCR primer sequence
}
}
\examples{
filename <- system.file("example_locus_attrs.csv", package = "chiimp")
locus_attrs <- load_locus_attrs(filename)
}
