% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_locus_attrs}
\alias{load_locus_attrs}
\title{Load table of locus attributes}
\usage{
load_locus_attrs(fp.locus_attrs, ...)
}
\arguments{
\item{fp.locus_attrs}{path to text file.}

\item{...}{additional arguments passed to \code{\link[utils]{read.table}}.}
}
\value{
data frame of locus attributes
}
\description{
Load a comma-separated table of locus attributes to use for analysis.
}
