% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsat.R
\name{logmsg}
\alias{logmsg}
\title{Write Log Message}
\usage{
logmsg(msg, end = "\\n")
}
\arguments{
\item{msg}{text to print.}

\item{end}{ending to concatenate to message; defaults to newline character.}
}
\description{
Print a log message to the standard error stream.
}
