% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsat.R
\name{main}
\alias{main}
\title{Handle full microsatellite analysis from command-line}
\usage{
main(args = NULL)
}
\arguments{
\item{args}{optional character vector of arguments to use rather than those
detected with \code{commandArgs}.}
}
\value{
list of results, with the full configuration list included as
  "config."
}
\description{
Read configuration from command-line arguments and run \code{full_analysis}.
}
