% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{make_allele_name}
\alias{make_allele_name}
\title{Create Short Allele Names}
\usage{
make_allele_name(data, hash_len = 6)
}
\arguments{
\item{data}{character vector of sequences}

\item{hash_len}{number of characters of alphanumeric hash to include as a
suffix.  (See \code{\link[openssl]{md5}}.)}
}
\value{
vector of short names
}
\description{
Autogenerate short names for sequences using sequence length and content.
}
