% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_cts_per_locus}
\alias{plot_cts_per_locus}
\title{Plot heatmap of read counts matching each locus primer}
\usage{
plot_cts_per_locus(cts_per_locus, idx.row = NULL, ...)
}
\arguments{
\item{cts_per_locus}{data frame as produced by
\code{\link{tally_cts_per_locus}}.}

\item{idx.row}{Optional vector of sample row indices to use.  (Using this
argument rather than filtering the input allows the same plot scale to be
used across plots.)}

\item{...}{additional arguments passed to \code{\link[pheatmap]{pheatmap}}.}
}
\description{
Given a data frame as produced by \code{\link{tally_cts_per_locus}}, plot a
heatmap showing the number of reads matching the forward primer of each locus
across samples.  Samples are shown on rows with the reads categorized by
locus across columns.
}
\seealso{
\code{\link{plot_heatmap}}
}
