% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_dist_mat}
\alias{plot_dist_mat}
\title{Plot Distance Matrix}
\usage{
plot_dist_mat(dist_mat, num.alleles = max(dist_mat),
  dist.display_thresh = round(num.alleles * 2/3), ...)
}
\arguments{
\item{dist_mat}{distance matrix as produced by
\code{\link{summarize_dataset}} via \code{\link{make_dist_mat}}.}

\item{num.alleles}{the maximum number of matching/mis-matching alleles.  Used
to determine color scaling.  Defaults to the highest observed distance in
the matrix.}

\item{dist.display_thresh}{distance value at or below which distances will be
explicitly drawn on the heatmap.  Above this value only the color-coding 
will signify distance.  Use \code{NA} to always show numbers.}

\item{...}{additional arguments passed to \code{\link[pheatmap]{pheatmap}}.}
}
\description{
Plot a heatmap of a distance matrix.
}
\seealso{
\code{\link{make_dist_mat}}
}
