% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_heatmap_stutter}
\alias{plot_heatmap_stutter}
\title{Plot heatmap of suspected PCR stutter}
\usage{
plot_heatmap_stutter(results_summary, ...)
}
\arguments{
\item{results_summary}{cross-sample summary data frame as produced by
\code{analyze_dataset}.}

\item{...}{additional arguments passed to \code{plot_heatmap}.}
}
\description{
Given a cross-sample summary data frame as produced by
\code{analyze_dataset}, plot a heatmap showing which samples had alleles
ignored due to suspected PCR stutter, with sample identifiers on rows and
loci on columns.
}
