% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{Extract Sample Attributes from Filenames}
\usage{
prepare_dataset(dp, pattern, ord = c(1, 2, 3), autorep = FALSE)
}
\arguments{
\item{dp}{directory path to search for matching data files.}

\item{pattern}{regular expression to use for parsing filenames.  There should
be exactly three groups in the pattern, for Replicate, Sample, and Locus.}

\item{ord}{integer vector giving order of the fields Replicate, Sample, and
Locus in filenames.  For example, if Locus is the first field followed by
Replicate and Sample, set \code{ord=c(3, 1, 2)}.}

\item{autorep}{logical allowing for automatic handling of any duplicates
found, labeling them as replicates.  FALSE by default.}
}
\value{
data frame of metadata for all files found
}
\description{
Find files matching a pattern in a given directory, and build a data frame of
standard sample attributes from fields in the filenames.  Alternatively, use
\code{\link{load_dataset}} to load a spreadsheet of sample attributes
explicitly.  \code{load_dataset} should be used for cases where more than one
locus is to be analyzed from a single sequencer sample (i.e., multiplexed
samples).
}
