% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primers.R
\name{raw_nt}
\alias{raw_nt}
\title{Make matrix of raw bytes from nucleotide sequences}
\usage{
raw_nt(seqs, map = RAW_NT, pad = 128, other = 0)
}
\arguments{
\item{seqs}{character vector of nucleotide sequences}

\item{map}{raw vector with nucleotide names and byte values}

\item{pad}{raw value to use for missing positions}

\item{other}{raw value to use for nucleotides not in \code{map}}
}
\value{
raw matrix with positions on rows and sequences on columns
}
\description{
Each sequence in the given vector becomes a column of the output matrix, with
a row for each position.  Shorter sequences are padded with a specific value
at bottom of the matrix.  With the defaults (see \link{RAW_NT}), A, C, G, and T
(case insensitive) are encoded as 01, 02, 03, and 04, IUPAC codes are bitwise
combinations of those values, padding values are 0x80, and any other
character is 0x00.
}
