% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiimp.R
\name{render_report}
\alias{render_report}
\title{Render Microsatellite Report Document}
\usage{
render_report(results)
}
\arguments{
\item{results}{list of microsatellite analysis results as produced by
\link{full_analysis}.}
}
\description{
Write an HTML report summarizing the results of a full microsatellite
analysis.
}
