% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_genotypes}
\alias{report_genotypes}
\title{Create genotype summary table}
\usage{
report_genotypes(results, na_replicates = "", na_alleles = "", closest = NULL)
}
\arguments{
\item{results}{list of results data as produced by \code{analyze_dataset}.}

\item{na_replicates}{text to replace NA entries with for the Replicates
column.}

\item{na_alleles}{text to replace NA entries with for the allele names}

\item{closest}{list of closest matches as produced by
\code{\link{find_closest_matches}}.}
}
\value{
data frame showing summary of genotypes.
}
\description{
Report the genotypes present in a processed dataset in a concise data frame.
This will arrange the allele names into a wide-format table with unique
samples on rows and loci on columns, do some automatic cleanup on the
columns, and show closest-matching individuals per entry, if given.  All NA
entries are replaced with blank strings or optionally (for NA Replicates or
untested sample/locus combinations) other custom placeholder text.
}
