% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_genotypes}
\alias{report_genotypes}
\title{Create genotype summary table}
\usage{
report_genotypes(tbl, allele.names = NULL, locus_attrs = NULL,
  hash.len = 6)
}
\arguments{
\item{allele.names}{optional data frame of custom allele names.  If present
the columns "Sequence" and "Name" will be used to map exact sequences to
custom names.}

\item{locus_attrs}{optional locus attributes data frame as produced by
\code{load_locus_attrs}.  Used to order locus columns in the returned data
frame.}

\item{hash.len}{number of characters of the sequence has to include when
auto-generating allele names.  If set to zero the hash suffix is not
included.}

\item{tbl.genotypes}{data frame as produced by \code{summarize_genotypes}.}
}
\value{
data frame showing summary of genotypes.
}
\description{
Report the genotypes present in a processed dataset, converting sequences to
short names with optional custom naming.  If custom names are not given or
names for sequences are not present in the custom names, a name is
autogenerated from the sequence length and a truncated hash.
}
