% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_idents}
\alias{report_idents}
\title{Create identification summary table}
\usage{
report_idents(results, closest, na_replicates = "")
}
\arguments{
\item{results}{results list as produced by \code{\link{summarize_dataset}}.}

\item{closest}{list of closest matches as produced by
\code{\link{find_closest_matches}}.}

\item{na_replicates}{text to replace NA entries with for the Replicates
column.}
}
\value{
data frame showing summary of sample genotypes with interleaved
  genotypes for similar known individuals.
}
\description{
Report the genotypes present in a processed dataset, paired with close
matches to known individuals, converting sequences to short names.  This is a
more detailed view than given by \code{\link{report_genotypes}}.
}
