% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_sample.R
\docType{data}
\name{sample_summary_funcs}
\alias{sample_summary_funcs}
\title{Versions of summarize_sample}
\format{
An object of class \code{character} of length 2.
}
\usage{
sample_summary_funcs
}
\description{
This list shows the different versions of \link{summarize_sample} recognized for
use via configuration in full_analysis.  See also \link{sample_analysis_funcs} and
\link{CFG_DEFAULTS}.
}
\keyword{datasets}
