% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_dataset}
\alias{save_dataset}
\title{Save table of sample attributes}
\usage{
save_dataset(data, fp, ...)
}
\arguments{
\item{data}{data frame of sample attributes as produced by \link{prepare_dataset}
or \link{load_dataset}.}

\item{fp}{path to text file.}

\item{...}{additional arguments passed to \link{save_csv}.}
}
\description{
Save a comma-separated table of sample attributes.  (This is a convenience
function not used automatically in the analysis.)
}
