% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{str_hist_render}
\alias{str_hist_render}
\title{Draw prepared histogram data}
\usage{
str_hist_render(bars, main, xlim, cutoff_fraction)
}
\arguments{
\item{bars}{list of data frames of counts-vs-lengths as prepared by
\code{\link{str_hist_setup}}.}

\item{main}{title of the plot.}

\item{xlim}{numeric range for x-axis.}

\item{cutoff_fraction}{numeric threshold for the fraction of locus-matching
counts needed to call an allele.  Used to draw a horizontal line if
\code{sample_data} is given.}
}
\description{
Render prepared histogram data to the display device.
}
