% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{str_hist_setup}
\alias{str_hist_setup}
\title{Prepare histogram data}
\usage{
str_hist_setup(seq_data, sample_data = NULL)
}
\arguments{
\item{seq_data}{data frame of dereplicated sequences as created by
\code{\link{analyze_seqs}}.}

\item{sample_data}{data frame of filtered and categorized sequences as
created by \code{\link{analyze_sample}}.}
}
\value{
list of data frames for the sets of counts-versus-length bars drawn
  in the plot, split by category.
}
\description{
Prepare histogram data
}
