% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_sample.R
\name{summarize_sample}
\alias{summarize_sample}
\title{Summarize a processed STR sample}
\usage{
summarize_sample(sample.data, locus.name, fraction.min, counts.min)
}
\arguments{
\item{sample.data}{data frame of processed data for sample as produced by
\code{analyze_sample}}

\item{locus.name}{character name of locus to summarize with}

\item{fraction.min}{numeric threshold for the minimum proportion of counts a
given entry must have, compared to the total matching all criteria for that
locus, to be considered as a potential allele.}

\item{counts.min}{numeric threshold for the minimum number of counts that
must be present, in total across entries passing all filters, for potential
alleles to be considered.}
}
\value{
list of attributes describing the sample.
}
\description{
Converts a full STR sample data frame into a concise list of consistent
attributes, suitable for binding together across samples for a dataset.  At
this stage the summary is prepared for a single specific locus, in contrast
to \code{analyze_sample}.
}
\details{
Entries in the returned list:
\itemize{
\item For Allele1 and Allele2:
\itemize{
\item Seq: sequence text for each allele.
\item Count: integer count of occrrences of this exact sequence.
\item Length: integer sequence length.
}
\item Homozygous: If the sample appears homozygous (if so, the Allele2 entries
will be NA).
\item Stutter: If a potential allele was ignored due to apparent PCR stutter.
\item CountTotal: The total number of sequences in the original sample data.
\item CountLocus: The number of sequences matching all criteria for the
specified locus in the original sample data.
\item ProminentSeqs: The number of entries above the specified threshold after
all filtering.  This should be either one (for a homozygous sample) or two
(for a heterozygous sample) but conditions such as cross-sample
contamination or excessive PCR stutter can lead to more than two.
}
}
