% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_sample.R
\name{summarize_sample_by_length}
\alias{summarize_sample_by_length}
\title{Summarize a processed STR sample, Length Version}
\usage{
summarize_sample_by_length(sample.data, sample.attrs, fraction.min = 0.15,
  counts.min = 500)
}
\arguments{
\item{sample.data}{data frame of processed data for sample as produced by
\code{\link{analyze_seqs}}.}

\item{sample.attrs}{list of sample attributes, such as the rows produced by
\code{\link{prepare_dataset}}.  Used to select the locus name to filter on.}

\item{fraction.min}{numeric threshold for the minimum proportion of counts a
given entry must have, compared to the total matching all criteria for that
locus, to be considered as a potential allele.}

\item{counts.min}{numeric threshold for the minimum number of counts that
must be present, in total across entries passing all filters, for potential
alleles to be considered.}
}
\value{
list of attributes describing the sample.
}
\description{
Summarize as in \code{\link{summarize_sample}}, but group entries by sequence
length rather than identity.
}
