% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_dataset.R
\name{tidy_analyzed_dataset}
\alias{tidy_analyzed_dataset}
\title{Tidy raw analyzed dataset results}
\usage{
tidy_analyzed_dataset(dataset, raw.results)
}
\arguments{
\item{dataset}{data frame of sample details as produced by
\code{\link{prepare_dataset}}.}

\item{raw.results}{list of pairs of sample summary and sample data data
frames (from \code{\link{summarize_sample}} and
\code{\link{analyze_sample}}).}
}
\value{
list of results, with \code{summary} set to the single summary data
  frame and \code{data} the per-sample data frames.
}
\description{
Rearrange pairs of sample summary / sample data objects into a single
summary cross-sample data frame and a list of detailed per-sample data
frames.
}
